1261 updagre get filtered sc output events for v27 (#1262)
* fix variable typo

* fix order of event parameter

* typos

* change op type id type

* update max gas to v27 values

+++ pkg/node/sendoperation/callsc/callsc.go
@@ -10,7 +10,7 @@ import (
 )
 
 const (
-	OpType = uint64(4)
+	OpType = 4
 )
 
 type OperationDetails struct {
